+++ migrations/lifecycle/CheckpointsFacet@0.1.0_deploy.js
@@ -1,13 +1,11 @@
-const {deployment} = require('../../src/templates/contracts');
-const {getContractAddress} = require('../../src/helpers');
+const Contract_deploy = require('../../src/templates/Contract/deploy');
+const {getContractAddress} = require('../../src/helpers/templates');
 
-module.exports = deployment(
-  'CheckpointsFacet@0.1.0',
-  'CheckpointsFacet',
-  [{name: 'ForwarderRegistry', value: getContractAddress('ForwarderRegistry@0.1.0')}],
-  {
-    importPath: 'node_modules/@animoca/ethereum-contracts-0.1.0/artifacts',
-    deterministicDeployment: true,
-    dependencies: ['ForwarderRegistry@0.1.0_deploy'],
-  }
-);
+module.exports = Contract_deploy('CheckpointsFacet@0.1.0', {
+  contract: 'CheckpointsFacet',
+  importPath: 'node_modules/@animoca/ethereum-contracts-0.1.0/artifacts',
+  args: [{name: 'ForwarderRegistry', value: getContractAddress('ForwarderRegistry@0.1.0')}],
+  deterministicDeployment: true,
+});
+module.exports.tags = ['Lifecycle'];
+module.exports.dependencies = ['ForwarderRegistry@0.1.0_deploy'];