+++ migrations/metatx/ForwarderRegistry@0.1.0_deploy.js
@@ -1,6 +1,8 @@
-const {deployment} = require('../../src/templates/contracts');
+const Contract_deploy = require('../../src/templates/Contract/deploy');
 
-module.exports = deployment('ForwarderRegistry@0.1.0', 'ForwarderRegistry', [], {
+module.exports = Contract_deploy('ForwarderRegistry@0.1.0', {
+  contract: 'ForwarderRegistry',
   importPath: 'node_modules/@animoca/ethereum-contracts-0.1.0/artifacts',
   deterministicDeployment: true,
 });
+module.exports.tags = ['Metatx'];