+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 - Used deterministic deployments in migrations.
 - Updated dependency version `@animoca/ethereum-contracts@0.1.0`.
-- Added migrations for `ForwarderRegistryContextFacet` and `ERC721`-related facets.
+- Added support for `ERC20` polygon predicates.
+- Added migration for `ForwarderRegistryContextFacet`.
+- Added migration and helpers for `MultiStaticCall`.
+- Reorganized helpers.
+- Added migrations for `ERC721`-related facets.
+- Added migrations for `ERC1155`-related facets.
 - Fixed bug in `skipIfFileExists`.
 
 ## 0.0.1